rename webappbase to product
