+++ pgml-dashboard/src/components/layouts/product/index/index.scss
@@ -1,0 +1,1 @@
+div[data-controller="layouts-product-index"] {}
