+++ pgml-dashboard/src/components/layouts/product/mod.rs
@@ -1,0 +1,6 @@
+// This file is automatically generated.
+// You shouldn't modify it manually.
+
+// src/components/layouts/product/index
+pub mod index;
+pub use index::Index;
