+++ pgml-dashboard/src/lib.rs
@@ -26,14 +26,15 @@ use guards::Cluster;
 use responses::{Error, Response, ResponseOk};
 use templates::{components::StaticNav, *};
 
+use crate::components::layouts::product::Index as Product;
 use crate::components::tables::serverless_models::{ServerlessModels, ServerlessModelsTurbo};
 use crate::components::tables::serverless_pricing::{ServerlessPricing, ServerlessPricingTurbo};
 use crate::utils::cookies::{NotificationCookie, Notifications};
 use crate::utils::urls;
 use chrono;
+use pgml_components::Component;
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
-use pgml_components::Component;
 
 #[derive(Debug, Default, Clone)]
 pub struct ClustersSettings {
@@ -325,7 +326,7 @@ pub async fn dashboard(tab: Option<&str>, id: Option<i64>) -> Redirect {
 
 #[get("/playground")]
 pub async fn playground(cluster: &Cluster) -> Result<ResponseOk, Error> {
-    let mut layout = crate::templates::WebAppBase::new("Playground", &cluster);
+    let mut layout = Product::new("Playground", &cluster);
     Ok(ResponseOk(layout.render(templates::Playground {})))
 }
