+++ pgml-dashboard/static/css/modules.scss
@@ -40,6 +40,7 @@
 @import "../../src/components/layouts/marketing/base/base.scss";
 @import "../../src/components/layouts/marketing/sections/three_column/card/card.scss";
 @import "../../src/components/layouts/marketing/sections/three_column/index/index.scss";
+@import "../../src/components/layouts/product/index/index.scss";
 @import "../../src/components/left_nav_menu/left_nav_menu.scss";
 @import "../../src/components/loading/dots/dots.scss";
 @import "../../src/components/loading/message/message.scss";
